+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.3] - 2025-07-14
+
+### Added
+
+- `is_allocated()` method for all memory map types to check allocation status of specific indices
+- `reset()` method for all memory map types to clear all allocations and return to initial state
+
 ## [v0.1.2] - 2025-07-04
 
 ### Changed